backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
pandas>=2.2,<3.0
django-cors-headers>=4.3,<5.0
reportlab>=4.0,<5.0
